# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OTMNow Delivery Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List resource allocations
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: resourceId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Create an allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Created allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
        segment:
          type: string
          enum: [enterprise, mid_market, smb]
        status:
          type: string
          enum: [active, inactive, prospect, at_risk, churned]
        healthScore:
          type: integer
        annualContractValue:
          type: number
        accountOwnerId:
          type: integer
        region:
          type: string
        otmVersion:
          type: string
        cloudDeployment:
          type: boolean
        createdAt:
          type: string
        renewalDate:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        segment:
          type: string
        status:
          type: string
        healthScore:
          type: integer
        annualContractValue:
          type: number
        accountOwnerId:
          type: integer
        region:
          type: string
        otmVersion:
          type: string
        cloudDeployment:
          type: boolean
        renewalDate:
          type: string
      required:
        - name
        - status
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        segment:
          type: string
        status:
          type: string
        healthScore:
          type: integer
        annualContractValue:
          type: number
        accountOwnerId:
          type: integer
        region:
          type: string
        otmVersion:
          type: string
        cloudDeployment:
          type: boolean
        renewalDate:
          type: string
...
    Allocation:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        resourceId:
          type: integer
        resourceName:
          type: string
        role:
          type: string
        allocationPct:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [confirmed, soft_booked, tentative, ended]
        hoursPerWeek:
          type: number
        createdAt:
          type: string
      required:
        - id
        - projectId
        - resourceId
        - allocationPct
        - status
        - createdAt
    CreateAllocationBody:
      type: object
      properties:
        projectId:
          type: integer
        resourceId:
          type: integer
        role:
          type: string
        allocationPct:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
        hoursPerWeek:
          type: number
      required:
        - projectId
        - resourceId
        - allocationPct
...